pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "seven-steps-to-poem"
version = "1.0.0"
description = "AI Agent system for automated business problem solving using McKinsey's 7-step methodology"
authors = ["Seven Steps Team <team@sevenstepstopoem.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "seven_steps"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.23"
asyncpg = "^0.29.0"
alembic = "^1.13.0"
redis = {extras = ["hiredis"], version = "^5.0.1"}
neo4j = "^5.15.0"
openai = "^1.3.0"
anthropic = "^0.7.0"
tenacity = "^8.2.3"
httpx = "^0.25.0"
celery = "^5.3.4"
temporal-client = "^1.0.0"
prometheus-client = "^0.19.0"
structlog = "^23.2.0"
typer = "^0.9.0"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sentence-transformers = "^2.2.2"
pandas = "^2.1.3"
numpy = "^1.25.2"
matplotlib = "^3.8.2"
seaborn = "^0.13.0"
scikit-learn = "^1.3.2"
networkx = "^3.2.1"
python-pptx = "^0.6.23"
pypdf2 = "^3.0.1"
openpyxl = "^3.1.2"
python-docx = "^1.1.0"
jinja2 = "^3.1.2"
boto3 = "^1.34.0"
minio = "^7.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"
pre-commit = "^3.6.0"
pytest-mock = "^3.12.0"
httpx-mock = "^0.10.0"
factory-boy = "^3.3.0"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.5.3"
mkdocs-material = "^9.4.0"
mkdocstrings = {extras = ["python"], version = "^0.24.0"}

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=seven_steps --cov-report=term-missing --cov-report=html"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["seven_steps"]
omit = [
    "*/tests/*",
    "*/venv/*",
    "*/__pycache__/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError"
]